Add cpedge protos for generic media stream messages (#4285)
Summary: We want to start streaming generic media streams in the UI. We
currently support just video streams. Upon further investigation, the
design of our device streaming mechanism is not video-specific. This PR
adds new protos for generic media stream stop/start/control.
For the edge -> CP messages, we also already handle this in a very
generic way. The device stream handler in the control plane doesn't look
at the contents of these messages and merely marshals it to JSON and
passes it to the websocket client. Following this, I decided not to add
a wrapper message around the edge -> CP message. adding that extra layer
doesn't seem to provide much benefit since we already depend on the
websocket client to parse the contents of the proto message anyway..

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: none, this just adds new protos that aren't used yet.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: b31e0c1b88303dfd79dc4d55e4a6a9d92bb88923