[compiler] Implement import from proto for Track LP node (#4270)
Summary: Implement the import.cc pass to convert pbtxt into the MLIR
format for tracking. The only weird thing that happens here is that we
don't want to specify the metadata path in the pbtxt, but we probably
want it in the pipeline definition? maybe there's another fix we need to
apply here.

Relevant Issues: Towards GML-1221

Type of change: /kind feature

Test Plan: Added a test to ensure the MLIR is as expected.

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 7e4b87178859057a2f419719ef71de10de9e30c7