[build] Fix opencv build (#4295)
Summary: The OpenCV build was failing on systems with `libjpeg-dev`
installed on the host. This PR forces OpenCV's cmake setup to build
libjpeg and libtiff from scratch instead of using host libraries.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested an uncached build of libjpeg on a machine with
`libjpeg-dev` installed.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: f52d14905d3199f2f4efa93460c6661353cae69c