[gem] New DetectionsMetricsSink calculator (#4292)
Summary: This is a new DetectionsMetricsSinkCalculator. It will replace
DetectionsSummaryCalculator, which is deprecated. We keep the old one
around to avoid breaking support, but it will be removed in the near
future.

Relevant Issues: Towards GML-1229

Type of change: /kind feature

Test Plan: Test case added.

Changelog Message: New DetectionsMetricsSinkCalculator in the GEM;
DetectionsSummaryCalculator is deprecated.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: a13a7cea24cd9b1cba108fce079b20cea6b620bb