Update device stream server to subscribe/publish on media stream topic (#4296)
Summary: The device stream component is already designed such that it is
not actually specific to the video stream. We just need to update it to
read/write to the new media stream instead.
However, for backwards compatibility with older GEMs, this keeps the
current path for videos. Once all the GEMs have been updated to handle
the new media stream message, we can completely remove that old path.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: skaffold an older GEM, check that the video stream still
works. skaffold a GEM updated to use the new media stream and switch to
the new topic, check the video stream still works.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 811a28d684acecf81f8faf98f662a0ea4ad14bfd