Add TextBatch proto for sending LLM outputs to media stream (#4299)
Summary: We want to be able to stream the results of an LLM to the UI.
This PR adds a proto which we can use to send the model's partial
results. It currently encodes the tokens as a string.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: none, this just adds a proto but doesn't use it yet.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 70184ad7a2037853906816c8bd3222b9234c309d