[gem calculator] Implement ScoresToClassificationCalculator (#4307)
Summary: Classification Metrics Sink compiler pass is broken because it
assumes that the ScoreToClassification calculator concretely exists, but
it is currently a temporary abstract op. James has had a long-running
TODO to actually complete this. This is step 1 where we split the first
half. This is what will get consumed by the downstream
classify metrics sink op.

Relevant Issues: n/a

Type of change: /kind feature

Test Plan: Added a test with a few cases in it.

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 98b6b4755073fad2f1b2e325f2ae82320d697bf1