Add media stream handler in GEM (#4297)
Summary: This adds the media stream handler to the GEM, which handles
any messages on the new media stream topic. This is basically copied
from the video stream handler, which we will remove once we've
completely deprecated that path.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: run GEM and check that the video stream works in the UI

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 39c2977e2ae17d4fd76155e1cb336268fe5641e4