[GEM] Add control message queue in execution context (#4315)
Summary: we need a way to pass the media stream control messages to our
pipeline. this adds a queue to the execution context which we add to as
we receive control messages. I considered using mediapipe's input
streams here, but decided against it since it seems more complicated
with the input naming in the graph.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: run GEM and publish control messages, check that the queue is
populated

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 2a05bc0c76a55a383a2daa0cb7bde870215889d3