[common] Add ArrayRef wrapper similar to llvm::ArrayRef (#4365)
Summary: Adds an `ArrayRef` object that can be used to keep a reference
to a contiguous array of elements without needing a specific container
type. This is mostly copied from LLVM's ArrayRef and is licensed as
such.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Added a simple unit test.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 6d7cfb11c68c1a70ba41a77a2f75b3831c41fba1