Add calculator for getting input from textStream (#4324)
Summary: This adds a calculator which reads from the mediastream control
queue and sends that as its output.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: attempted to add a unit test, however it seems the mediapipe
runner runs processImpl in an infinite loop if there are no expected
inputs. I debated adding an input just for testing, which worked, but
seemed hacky. The other test (opencv_cam_calculator_test) for a
calculator without inputs takes a max_num_frames which ends the
execution.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 7a94b26cf62cf821abc765828d3803b01812d304