[gem] ClockLatencyMetricsSinkCalculator (#4373)
Summary: To support proper recording of latencies, the compiler needs to
switch from the PacketLatencyCalculator--which doesn't work in loop
regions--to a combination of ClockTimestampCalculators and
ClockLatencyCalculators. The GEM then will then need to consume the data
from the ClockLatencyCalculator into this
ClockLatencyMetricsSinkCalculator to record the latencies. This PR adds
the GEM component of this change.

Relevant Issues: Towards GML-1278

Type of change: /kind feature

Test Plan: Test case added. Also tested end-to-end with the compiler
changes that are part of an upcoming PR.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: 342e72957fffcb753099d3570341366398730c89