Add gRPC qdrant client (#4370)
Summary: This will be used in the qdrant calculator to search for
documents given an embedding. I didn't include the clientrunner binary
which I used to test the client, since I was just using that for my own
testing.

Relevant Issues: towards GML-1245

Type of change: /kind feature

Test Plan: use the client and make sure it connects and runs query

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: e3e7ee7b434a734c0fe8a9c80500d64f2ed0dfe9