[compiler] Support `Opaque` models (#4407)
Summary: This adds support for "opaque" model storage format. These
kinds of models will be entirely ignored by the compiler and passed
through to the GEM without attempting to import to MLIR or convert the
model. This will be used for the huggingface tokenizer model files.

Relevant Issues: Towards GML-1270

Type of change: /kind feature

Test Plan: Tested locally that the opaque models are ignored by the
compiler.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 23f78ff162571a3d2eb79c5e42c52326efe84b44