[pyapi] Change collect_assets to a context manager (#4429)
Summary: This PR changes `collect_assets` to be a context manager. This
will allow `Model` subclasses to handle lifecycles of their own assets.
For example, to use a temporary file that gets deleted after the call to
`collect_assets`.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: N/A

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: eeb7490db7bb905156ed3dd83917c2847794f0be