[gem] Update name of classification metric (#4434)
Summary: Change the classification metric name from "confidence" to
"scores" which more accurately represents what it is. To avoid breaking
old control planes, we keep the old metric name as a separate metric. We
can remove it in the near future.

Relevant Issues: N/A

Type of change: /kind bugfix

Test Plan: Updated the test.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: 1f7c844d512498da2a69e896cd1a91a73ed5c2f8