[pyapi] Allow skipping model uploads (#4444)
Summary: This PR changes the behaviour of `upload_pipeline` such that it
skips uploading models that already exist. Previously, `upload_pipeline`
would error out if any of the models already existed. Now it will just
print a warning telling the user that the model was not uploaded since
it already exists.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested that model upload still works and that if the model
already exists `upload_pipeline` can still succeed.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 7faebb0469a4713f70b60073395a3a62fffffd76