[pyapi] Add AttentionKeyValueCache TensorSemantics (#4527)
Summary: Adds tensor semantics for `AttentionKeyValueCache` which are
tensors that store the key value cache for a transformer, and should be
passed through from the output of a previous step to the inputs of the
next step.

Relevant Issues: Towards GML-1260.

Type of change: /kind feature

Test Plan: Tested end to end.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 8aa4d55d2dad1c57f253c0ba164349401c92f884