[gem] Add stateful loopback in OpenVINOExecuteCalculator (#4526)
Summary: Adds calculator options to `OpenVINOExecuteCalculator` to allow
looping back outputs from the previous timestamp into inputs for the
next timestamp. The current use case for this is LLM Key/Value caches.
As part of this an optional input is added to the calculator to force a
reset of the loopback.

Relevant Issues: Towards GML-1260.

Type of change: /kind feature

Test Plan: Tested deploying YOLO to make sure the non-loopback stuff
still works. Also tested with an LLM to ensure the loopback works
properly.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 21fed7d9466ae7d59bb1953eea2a42cb2880f7d4