[pyapi] Add key-value caching to huggingface text generation pipelines (#4528)
Summary: Adds a key-value cache when tracing huggingface text generation
pipelines. This adds the cache in an explicit way where each iteration
the cache is passed in as an input and then the new cache is output.

Relevant Issues: Towards GML-1260.

Type of change: /kind feature.

Test Plan: Tested with both SmolLM and LLama3.1

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 6a7b1c3b8554c8fa7ceec7a34e42b714a2857be7