[pyapi] Pin the torch-mlir-gml version since it no longer is specific to the pytorch version. (#4590)
Summary: The new version of torch-mlir-gml I just released, no longer
relies on a specific version of torch. So I reset the version numbers on
torch-mlir-gml, and I'm pinning it to 0.0.1 in our python API. This new
version works with torch >=2.3 so I added that requirement as well.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested that importing models still works with the new
version.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 099aa39e71bde0a60fd1318738a40b9d182b7490