Update qdrant calculator with reconnection logic (#4587)
Summary: We only start the gRPC connection on `Open` currently. This can
lead to issues if the connection has been opened for a long amount of
time (timeouts, server disconnects). This adds logic to the calculator
so that if the channel is ever closed, we reconnect.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: run GEM

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: da30a2341777f8f800ee48e6bd981d35743c0736