[gem] Clamp latency metrics to be minimum 0 (#4580)
Summary: Because of mediapipe scheduling, the latency metrics can be
negative for our implementation of a LatencyMetricsSink. There are some
fundamental problems that we need to address to figure this out, they
would require a large refactor. This is not something we want to do
given that the inaccuracies aren't that large.
This issue tracks the possible solutions: [GML-1320: Address the cause
of latency calculator metrics
inaccuracy](https://linear.app/gimletlabs/issue/GML-1320/address-the-cause-of-latency-calculator-metrics-inaccuracy)

Relevant Issues: n/a

Type of change: /kind bugfix

Test Plan: Added a test

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: eb775535ad9e721d81c1030c0a2fbfa67b883ed4