[gem] Record time to last token in otel histogram (#4591)
Summary: Record the time to last token. Yesterday we tried to add
latency calculators for the GenerateTokens call, but ran into an issue
where the "end point" of the recording is not well defined. We decided
for now to record the data to a different metric and merge the data in
the UI instead of trying to fitting
the final token output into the latency metrics.

Relevant Issues: n/a

Type of change: /kind feature

Test Plan: Added tests

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 9319b697f4d62b8f06228ebdb75b877e757a3694