[reconciliation] Add device resource versions to physicalpipeline spec/status (#4600)
Summary: We currently have an issue where if a device's capabilities
change, we don't trigger a recompile. This is because we currently don't
have any information for whether a device's capabilities have changed or
not.
This adds deviceResourceVersion to the spec/state protos, so we can
track what resourceVersion those spec/states are for. This also adds a
util for getting a deviceResourceVersion string given a device's
capabilities.
This does not use it yet in the reconciliation flow.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: unit test

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: d36991e4a12eda842cb1231bca72ca6aa99ec852