[gem] Clean-up calculator tests to use test infrastructure (#4631)
Summary: Switch a number of our tests to use the CheckMetrics helper,
which simplifies the code a bunch. Also adjust CheckMetrics so that it
works with multiple metrics per calculator. Note that this PR also
removes a deprecated metric.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: The tests themselves.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: aef0c06525104a832d8ff3ff5498432c7949dd35