Push sync_set handler options into the calculator contract (#4654)
Summary: Our LLM token calculators expect inputs to be in certain
`sync_sets` and aren't designed to work with the default input handler.
So we can just push the input handler config into the contract and
remove the need for the calculator consumer/creator to have to specify
the handler options.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: The calculator unit tests continue to work with or without
the input handler options explicitly specified.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 2934c08656d93b13b66ab801031c3004405b6cbb