[gem] Remove lost tracks metric from the tracks_metrics_sink_calculator (#4664)
Summary: Lost tracks metrics don't work right now. We don't plan to
allocate time to fix this anytime soon, we may revisit recording this
data in the future if users request it, but in the meantime we just stop
recording a metric so that users
don't get the idea that they can actually query this data.

Relevant Issues: Closes GML-1311

Type of change: /kind cleanup

Test Plan: Updated the test

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 6bf71f15141237cb6f84139d46c6b42f95a22f3b