Patch torch so that it finds all nvidia libs (#4675)
Summary: This is a bug upstream. This doesn't trigger for torch installs
outside of bazel becuase the LD search paths are setup to expect the
nvlibs pypi packages to be installed in a particular relative location
compared to torch. When installing with bazel, these abstractions break
down and we rely on the sys path based lib searching that pytorch has.
See https://github.com/pytorch/pytorch/issues/131284 and
https://github.com/pytorch/pytorch/pull/131295

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Ran the tests that depend on torch, they now pass in the
sysroot.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: dbd181e28d2a46f27e8bea1f0f9e37803c6a14ce