[gem] MergeTokensCalculator: Add output for prompt timestamp (#4668)
Summary: To measure latency properly in graphs, we need to be able to
convert any internal timestamps back to their original domain. Since the
MergeTokensCalculator creates a new internal timestamp domain, output
the original timestamp domain as an output so that a mapping back to the
original can be made if desired. This will be used in an upcoming PR.

Relevant Issues: Towards GML-1339

Type of change: /kind feature

Test Plan: Test case updated, though it has limitations. Our current
testing infrastructure does not allow us to easily test a graph with
multiple different timestamp domains. But the coverage is probably
sufficient for now.

Changelog Message:

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: cb13f358eb3fa379331af475850a96c961d0059c