[gem] MergeTokensCalculatorTest: Restructure to enable different timestamp domains (#4680)
Summary: This PR makes the MergeTokensCalculatorTest more robust by more
accurately modeling the concept of timestamp domains. In this particular
calculator, the inputs are in one timestamp domain, while the tokens are
in a different timestamp domain that recirculate. By switching away from
CalculatorRunner, we can more finely control the sequence of events,
particularly through the use of WaitUntilIdle(), to more accurately
reflect how the calculator would be stimulated in a real graph.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: This is a test improvement.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: dac5b2513f2e313f8a7b554feb5064d52d2d6e69