[gem] TextStreamSinkCalculator: Generate output to mark completion in the prompt time domain (#4674)
Summary: The TextStreamSinkCalculator is the last calculator in the
token time domain, but for latency measurement purposes, we also need a
packet that is in the original graph's time domain. As such, this change
emits a new FINISHED packet from that corresponds to the original time
domain. The FINISHED signal is generated whenever an EOS packet is
received.

Relevant Issues: Towards GML-1339

Type of change: /kind feature

Test Plan: Updated the test to look for the new output, at the original
timestamp domain.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: c4c94d18fa42ed1200d245afb9b4e5ab18e41899