[build] Switch to using libstdc++ to fix exceptions in shared libs (#4701)
Summary: Prior to this change we were statically linking libc++ into our
binaries. However, this can cause issues when shared libraries are
loaded that link to `libstdc++` dynamically (likely an issue even if
they are using `libc++` but no one links that dynamically so we've never
seen that). In particular, we were seeing segfaults when exceptions are
thrown in shared library code. This became especially egregious in
python where everything is a shared lib, and nothing is statically
linked. To fix these issues, this PR switches our build to use libstdc++
by default instead of libc++. Also, it changes our libstdc++ build to
link stdc++ dynamically instead of statically. (I tried linking it
statically and it still causes the same issue albeit in fewer
situations).

I tried to use debian10 so that we could support ubuntu 18.04 as well,
but I ran into some issues with `std::filesystem` (there are segfaults
in destructors when using `std::filesystem` from shared libraries.)

Relevant Issues: N/A

Type of change: /kind build

Test Plan: Tested that both jetson and nucs still work with a
controlplane and gem built with libstdc++. Also tested with other
changes that the exception issue appears to be fixed.

---------

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 598982e7621854bea5773457b4ee74ef262f22dc