Update device capabilities to send up supported camera types (#4708)
Summary: This updates the device info protos to include info about all
supported camera types even if the device might not have a camera of
that type. This is to support muxing into cameras that are hot plugged
in.

Relevant Issues: Fixes GML-1348

Type of change: /kind feature

Test Plan: Existing build and tests pass. Deployed a dev gem with vlog
and checked that the proto includes the supported cameras.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: de7863076a13c8032b843be71a7417e99d8e8554