ui: Create a jsdom extended env that include globals that are in nodejs (#4720)
Summary: This is to work around the fact that the jsdom env is missing
some globals that are and should be available from nodejs (they are not
polyfilled nor exposed from the node env).

Relevant Issues: N/A

Type of change: /kind test

Test Plan: Tested the updated text with bazel and directly with jest.

---------

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 55912b17524b518c50ab279417647291bc952b08