[cleanup] Add RECORD patch for torch cpu wheel (#4729)
Summary: Adds a patch for the RECORD file in the cpu pytorch wheel,
similar to the patch for the cuda version. Unclear why this is necessary
but it silences bazel's debug warning about it.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Ran bazel build //src/api/python/... without seeing the
warning.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 79920ec54143a6ef32c3b05f928f1831c1d121cc