[gem] Support external weights in ONNX models for TensorRT (#4749)
Summary: The TensorRT ONNX parser already supports ONNX's EXTERNAL
location api for storing onnx weights in external files. However, it
requires that the files are all in the same folder and that the base
names of the files are the same as in the TensorProto in the ONNX file.
So we need to build a temp dir that contains each of the model assets.
We use the asset names as the file names so that the compiler can
reference the asset name in the ONNX file itself.

Relevant Issues: Towards GML-1332.

Type of change: /kind feature

Test Plan: Tested end to end, saw that a model with external weights
works as expected.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: f370e5698484d63041753c55ec859da5a6842408