[compiler] Handle TensorReferenceOp in openvino export (#4759)
Summary: Handle the new safetensors.TensorReferenceOp during export to
openvino. Since openvino only supports a single weight file, the
exporter has to calculate offsets based on the assumption that all the
shards will be concatenated in order. The alternative is to have the
exporter concatenate all the weights into a single file. The benefit of
doing it the sharded way, is that the input weight shards are used
unchanged so the compiler service doesn't need to re-upload those files.
This can be a huge benefit if the weight files are large.

Relevant Issues: Towards GML-1332.

Type of change: /kind feature

Test Plan: Added lit test for conversion of this op. Also tested end to
end.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: f4eae4b9f7f41635a15a29bf25755f2b208c3ed4