infra: Don't explicitly pull linter image when running lint (#4765)
Summary: `mega-linter-runner` tries to always pull the linter image
since they reuse the tag. We use an explicit tag for the image that
isn't overwritten so we don't need the always pull.
This is useful since it speeds up the linter run and also doesn't
require fresh gcloud credentials incase the developer is using the
gcloud auth plugin for other images.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Ran `make lint`. Saw that the pull was skipped. Noticed that
changing the tag caused docker to implicitly pull the image that was not
on my machine.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: ef2080fd30de413262a71badbbb737beed68fa84