[pyapi] Support storing weights in separate safetensor files (#4764)
Summary: This PR adds support to the python API to output weights to
safetensor files instead of inline in the MLIR.

Relevant Issues: Towards GML-1332.

Type of change: /kind feature

Test Plan: Tested that this works outside of bazel. A followup PR will
update the conversion tests to use this new way of importing.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 26c7abb10a8f0f59674d3321ba727f74fd88e719