Fix default calculator (#4794)
Summary: The default graphs that show the video were broken after #4776.
I haven't figured out why the graph level options want the proto from
the framework, but listing this as a dep is enough to fix it. Let's go
with this change until we figure out what's going on here.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Tested by skaffolding a gem and ensuring that the default
graph worked.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 882a2d0b0c68cd2b85d7d3438110e52c1e161227