Remove dependence on frameRate in ffmpeg encoder (#4800)
Summary: The framerate in the codec ctx struct is used to indicate the
rate for CFR streams. Since our graph can drop packets, the stream is no
longer CFR, so we can just omit this param.
The `time_base` is often set to `1/fps` to indicate the seconds between
frames (when the frame timestamps are incremented by 1). We have real
timestamps in microseconds so our timebase is 1us and not `1/fps`

Relevant Issues: Fixes GML-1403

Type of change: /kind cleanup

Test Plan: Tested with a skaffolded GEM.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: de3a80e292c54e5732912601dc3c5d35af96bf08