Add deviceConfig stream to track changes to device configs (#4793)
Summary: Messages will be put on this stream whenever a base config is
updated or device/fleet tags are updated. This will trigger a check to
whether a device's config is up-to-date, and send out that new config to
the device if not.

Relevant Issues: towards GML-1346

Type of change: /kind feature

Test Plan: none yet, this is currently unused

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: baac5fd543fcd071e42b7a527ef2aeccfa97b9bd