cleanup: Remove frame_rate from VideoHeader (#4808)
Summary: This removes the frame rate from the videoheader we use in the
MP graph and also from the data we send to the controlplane/UI.
Due to backpressure/packet thinning out framerate isn't constant nor
statically known so we shouldn't rely on it anyway.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Deployed a gem+controlplane and ensured that the pipelines
and video still work.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 428b9a41b264fe75d9ba0baf192a7e84b70c2184