bug: Check isOpened for opencv based camera sources (#4832)
Summary: When overriding the camera with a file, it's easy to make a
typo/mistake and then not realize what happened. This checks IsOpened
and errors if we failed to open the source.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Tested by running a GEM with a nonexistent video.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 0bda37825283a7a974bdc55b66a433b64e6e32bc