[controlplane] Add msgprocessor to handle BaseConfigUpdateRequest jetstream messages (#4866)
Summary: This adds the handler which subscribes to the DeviceConfig
stream and handles any `BaseConfigUpdateRequest` messages. When
receiving this message, we need to compare the current state of the
device config and compute the desired base config.
The logic for computing the desired base config is into a TODO, since we
don't yet have the contents of the GEMConfig filled out.
I also moved `loadTestData` to a shared testutils for the fleetmgr
service, since we use it in the msgprocessor + controller code.

Relevant Issues: towards GML-1346

Type of change: /kind feature

Test Plan: added unit test which checks if a NATS message is sent out in
response

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 10a6924bac5e23249b45bda0847233a1e80ace28