[pyapi] Add workaround for importing models that do not follow pytorch conventions. (#4850)
Summary: This adds a workaround for importing pytorch models that don't
follow conventions around submodule registration. In particular, this
workaround fixes most (all known) cases of the `module is not installed
as a submodule` error.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Added tests for a number of different cases of the `module is
not installed as a submodule`. Prior to these changes all those test
cases failed.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: f50da0a468e1411ae815c1052a08af5ba22751f7