[gem] Tweak log of mediapipe graph (#4879)
Summary: Move log to Init() instead of Start(), which gives a better
chance of it being printed out in case of errors during Init(). This log
also prints out the model spec info. Finally, the version printed here
is easier to use with https://viz.mediapipe.dev/ as it emits fewer
special characters that need to be removed.

Relevant Issues: N/A

Type of change: /kind internal

Test Plan: Manually checked the log.

GitOrigin-RevId: 9dab08628e6baf35a170f249e535dbf72b27665a